feat(base): group reth and snapshot commands under subcommand groups
Reorganize the unified `base` binary so reth-derived maintenance commands live under `base reth <db|init|init-state|dump-genesis|stage|p2p|config|prune|re-execute>` and snapshot commands live under `base snapshot <manifest|download>`. `base proofs` stays at the top level. This flattens the surface a bit while grouping related tooling and keeps `--chain` isolated on each reth subcommand.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
