+++ bin/base/src/commands/mod.rs
@@ -3,6 +3,8 @@
 mod bootnode;
 mod command;
 pub(crate) use command::BaseCommand;
+mod reth;
 mod rpc;
 mod sequencer;
+mod snapshot;
 mod update;
